# apps/target_app/stock_app/src/internal/apidoc/openapi.yaml
openapi: 3.0.3
info:
  title: stock_app API (Demo Target)
  version: 1.0.0
  description: |
    The production application observed by the **Incident Pilot SRE agent**.
    A modular monolith (user / stock / order) on CockroachDB, deployed as a
    single ECS Fargate container.

    **How to use this page:**
    1. `POST /api/v1/auth/register` (public), then `POST /api/v1/auth/login` — copy the `token`.
    2. Click the **Authorize** button, paste the token, confirm.
    3. Create products, reserve stock, and place orders (placing an order
       reserves stock in-process across the stock/order databases).
servers:
  - url: /
    description: This server (same origin as the docs page)
tags:
  - name: health
    description: Liveness and database reachability
  - name: auth
    description: Registration, login, and the current user
  - name: users
    description: User management
  - name: stock
    description: Products and stock reservation (stock_db)
  - name: orders
    description: Orders (order_db); placing an order reserves stock in-process

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [health]
      summary: Liveness + CockroachDB reachability
      description: Also feeds the DatabaseLatency / DatabaseUp metrics the SRE agent reads.
      security: []
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthOK' }
        '503':
          description: Database down
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a user (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Invalid body / validation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login and get a bearer token (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Token + user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: The user identified by the bearer token
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          description: Missing/invalid token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/users:
    get:
      tags: [users]
      summary: List users
      responses:
        '200':
          description: Users (newest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [users]
      summary: Create a user (no password)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Invalid body / validation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [users]
      summary: Get a user
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [users]
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [users]
      summary: Delete a user
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/stock/products:
    get:
      tags: [stock]
      summary: List products
      responses:
        '200':
          description: Products (newest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [stock]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '400':
          description: Invalid body / validation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/stock/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [stock]
      summary: Get a product
      responses:
        '200':
          description: The product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/stock/products/{id}/reserve:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [stock]
      summary: Reserve stock (decrement quantity)
      description: |
        Manual/testing entry point. The order context reserves in-process instead
        (POST /api/v1/orders) — this is the same underlying call.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveRequest' }
      responses:
        '200':
          description: Reserved; the unit price at reservation time
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReserveResponse' }
        '400':
          description: Invalid body / quantity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Product not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Insufficient stock
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/orders:
    get:
      tags: [orders]
      summary: List orders
      responses:
        '200':
          description: Orders (newest first, without items)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
    post:
      tags: [orders]
      summary: Place an order (reserves stock in-process)
      description: |
        Reserves every line against the stock service in-process (stock_db),
        then creates the order and its items in one transaction (order_db). If any
        line cannot be reserved, no order is created. `price_at_purchase` snapshots
        the unit price so later repricing never rewrites an existing order.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrderRequest' }
      responses:
        '201':
          description: Created (status PENDING)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Invalid body / validation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Insufficient stock, or a product was not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/v1/orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [orders]
      summary: Get an order (with its items)
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Paste the token from POST /api/v1/auth/login (24h TTL).'

  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        quantity: { type: integer }
        price: { type: number, format: double }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    OrderItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        quantity: { type: integer }
        price_at_purchase: { type: number, format: double }

    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status:
          type: string
          enum: [PENDING, PAID, SHIPPED, CANCELLED]
        total_price: { type: number, format: double }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    CreateUserRequest:
      type: object
      required: [username, email]
      properties:
        username: { type: string, minLength: 1 }
        email: { type: string, format: email }

    CreateProductRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0, default: 0 }
        price: { type: number, format: double, minimum: 0, default: 0 }

    ReserveRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1 }

    ReserveResponse:
      type: object
      properties:
        product_id: { type: string, format: uuid }
        unit_price: { type: number, format: double }

    PlaceOrderRequest:
      type: object
      required: [items]
      properties:
        user_id:
          type: string
          format: uuid
          description: Optional — defaults to the authenticated user from the token.
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderItemInput' }

    OrderItemInput:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1 }

    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }

    HealthOK:
      type: object
      properties:
        status: { type: string, example: ok }
        service: { type: string, example: stock_app }
        database: { type: string, example: up }

    Error:
      type: object
      properties:
        error: { type: string }